% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedi.R
\name{cedi}
\alias{cedi}
\title{A classical-CEDI Function}
\usage{
cedi(try = TRUE)
}
\arguments{
\item{try}{Defaults to TRUE.}
}
\value{
string of characters
}
\description{
This function allows to test my package.
}
\examples{
c <- test_function(TRUE), print(c) / c <-test_function(FALSE), print(c)
test_function()
}
\keyword{test}
