% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fun.R
\name{test_fun}
\alias{test_fun}
\title{A Test Function}
\usage{
test_fun(try = TRUE)
}
\arguments{
\item{try}{Ciao Zoe, sta bene? Defaults to TRUE.}
}
\value{
string of characters
}
\description{
This function allows to test my package.
}
\examples{
c <- test_function(TRUE), print(c) / c <-test_function(FALSE), print(c)
test_fun()
}
\keyword{test}
